Cleanup, added other database calls to support rating histories

+++ API/Services/Interfaces/IPlayerService.cs
@@ -5,5 +5,5 @@ namespace API.Services.Interfaces;
 public interface IPlayerService : IService<Player>
 {
 	Task<Player?> GetByOsuIdAsync(int osuId);
-	Task<int> GetIdByOsuIdAsync(int osuId);
+	Task<int> GetIdByOsuIdAsync(long osuId);
 }
